requirements.txt
pandas==2.1.4
numpy==1.26.4
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-dotenv==1.0.1
tqdm==4.66.2
loguru==0.7.2
pytest==8.0.0
